shoplifting_detection_system/requirements.txt
# Core Framework Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database Dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Machine Learning and Computer Vision
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.206
opencv-python==4.8.1.78
mediapipe==0.10.8
scikit-learn==1.3.2
numpy==1.24.4
scipy==1.11.4

# Image and Video Processing
Pillow==10.1.0
imageio==2.31.6
av==10.0.0

# Data Processing and Analysis
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

# Async and Concurrency
asyncio-mqtt==0.16.1
aiofiles==23.2.1
aioredis==2.0.1

# Monitoring and Metrics
psutil==5.9.6
GPUtil==1.4.0
prometheus-client==0.19.0

# Networking and APIs
requests==2.31.0
httpx==0.25.2
websockets==12.0

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1
configparser==6.0.0

# Security and Encryption
cryptography==41.0.8
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Notification Services
smtplib2==0.2.1
twilio==8.10.3
pyfcm==1.5.4

# Message Queues and Streaming
kafka-python==2.0.2
celery==5.3.4
redis==5.0.1

# Testing and Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3

# File Handling
openpyxl==3.1.2
xlsxwriter==3.1.9

# Logging and Debugging
loguru==0.7.2
sentry-sdk==1.38.0

# Performance Optimization
numba==0.58.1
cython==3.0.6

# Containerization Support
gunicorn==21.2.0
docker==6.1.3

# Development Tools
jupyter==1.0.0
ipython==8.18.1
notebook==7.0.6

# Additional ML Libraries
transformers==4.36.2
timm==0.9.12
albumentations==1.3.1

# Video Analytics
ffmpeg-python==0.2.0
moviepy==1.0.3

# Geospatial (if needed for camera positioning)
geopy==2.4.1

# API Documentation
fastapi-users==12.1.2
fastapi-pagination==0.12.13

# Caching
diskcache==5.6.3
cachetools==5.3.2

# Data Validation
marshmallow==3.20.1
cerberus==1.3.5

# Time Series Analysis
statsmodels==0.14.0

# Graph Processing (for relationship analysis)
networkx==3.2.1

# Memory Profiling
memory-profiler==0.61.0
pympler==0.9

# System Integration
systemd-python==235
supervisor==4.2.5

# Additional Computer Vision
dlib==19.24.2
face-recognition==1.3.0

# Advanced Analytics
plotly==5.17.0
dash==2.14.2

# Model Serving
onnx==1.15.0
onnxruntime==1.16.3

# Distributed Computing
ray==2.8.1

# Configuration Management
hydra-core==1.3.2
omegaconf==2.3.0

# Experiment Tracking
wandb==0.16.1
mlflow==2.8.1

# Data Pipeline
apache-airflow==2.7.3
prefect==2.14.11

# Advanced Monitoring
grafana-api==1.0.3
influxdb-client==1.38.0

# Security Scanning
bandit==1.7.5
safety==2.3.5

# Code Quality
pre-commit==3.6.0
isort==5.12.0

# Performance Testing
locust==2.17.0

# API Rate Limiting
slowapi==0.1.9

# Advanced Serialization
msgpack==1.0.7
orjson==3.9.10

# Backup and Recovery
boto3==1.34.0  # For AWS S3 backup
azure-storage-blob==12.19.0  # For Azure backup

# Advanced Networking
paramiko==3.4.0  # For SSH connections
fabric==3.2.2  # For remote deployment

# Database Migrations
yoyo-migrations==8.2.0

# Advanced Logging
structlog==23.2.0
python-json-logger==2.0.7

# Health Checks
healthcheck==1.3.3

# Process Management
supervisor==4.2.5
circus==0.18.0

# Advanced Configuration
dynaconf==3.2.4

# Model Optimization
torch-tensorrt==1.4.0
tensorrt==8.6.1

# Edge Deployment
onnxruntime-gpu==1.16.3
openvino==2023.2.0

# Advanced Metrics
statsd==4.0.1
datadog==0.47.0

# Load Balancing
haproxy-stats==2.4.22

# Service Discovery
consul-python==1.1.0
etcd3==0.12.0

# Advanced Security
authlib==1.2.1
oauthlib==3.2.2

# Data Encryption
fernet==1.0.1
keyring==24.3.0

# Advanced Caching
memcached==1.59
pymemcache==4.0.0

# Message Encryption
pynacl==1.5.0
cryptography==41.0.8

# Advanced File Handling
h5py==3.10.0
zarr==2.16.3

# Distributed Storage
minio==7.2.0
hdfs3==0.3.1

# Advanced Analytics
apache-superset==3.0.3
metabase-api==0.3.0

# Model Versioning
dvc==3.30.3
git-lfs==1.6

# Advanced Deployment
kubernetes==28.1.0
helm==3.13.2

# Service Mesh
istio==1.19.4

# Advanced Monitoring
jaeger-client==4.8.0
opentelemetry-api==1.21.0

# Load Testing
artillery==2.0.0
k6==0.47.0

# Advanced Security
vault==0.10.0
hashicorp-vault==1.0.0

# Container Security
trivy==0.47.0
clair==4.7.1

# Advanced Networking
nginx==1.25.3
traefik==3.0.0

# Advanced Database
clickhouse-driver==0.2.6
timescaledb==2.12.0

# Advanced Search
elasticsearch==8.11.0
opensearch-py==2.4.0

# Advanced Messaging
apache-pulsar==3.4.0
nats-py==2.6.0

# Advanced Streaming
apache-kafka==3.6.0
apache-storm==2.5.0

# Advanced ML Ops
kubeflow==1.8.0
mlflow==2.8.1

# Advanced Data Processing
apache-spark==3.5.0
dask==2023.11.0

# Advanced Visualization
bokeh==3.3.2
altair==5.2.0

# Advanced Testing
hypothesis==6.92.1
factory-boy==3.3.0

# Advanced Documentation
mkdocs==1.5.3
gitbook==4.0.0

# Advanced Deployment
ansible==8.7.0
terraform==1.6.5

# Advanced Monitoring
new-relic==9.2.0
datadog==0.47.0

# Advanced Security
snyk==1.1.1
sonarqube==1.0.4
